src-tauri/pyproject.toml
[project]
name = "tauri-app"  
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.9"
dependencies = [
    "pytauri == 0.8.*",
    "pydantic == 2.*",
    "anyio == 4.*",
    "SQLAlchemy >= 2, < 3",
    "openai >= 1.30.0",
    "python-dotenv >= 1.0.1"
]  

[project.entry-points.pytauri]
ext_mod = "tauri_app.ext_mod"

[build-system]
requires = ["setuptools >= 80"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages]
find = { where = ["python"] }
